Factor out feature quant specific ops
