+++ gigl/distributed/distributed_neighborloader.py
@@ -27,9 +27,9 @@ from gigl.distributed.sampler_options import (
 from gigl.distributed.utils.neighborloader import (
     DatasetSchema,
     SamplingClusterSetup,
-    append_dequantized_node_features,
     extract_metadata,
     labeled_to_homogeneous,
+    materialize_quantized_node_features,
     set_missing_features,
     shard_nodes_by_process,
     strip_label_edges,
@@ -555,7 +555,7 @@ class DistNeighborLoader(BaseDistLoader):
             data = labeled_to_homogeneous(DEFAULT_HOMOGENEOUS_EDGE_TYPE, data)
 
         data, metadata = self._apply_ppr_outputs(data, metadata)
-        data, metadata = append_dequantized_node_features(
+        data, metadata = materialize_quantized_node_features(
             data=data,
             metadata=metadata,
             node_quantization_metadata=self._node_quantization_metadata,
